+++ examples/CustomDensity/LiteMol-example.js
@@ -40,6 +40,7 @@ var LiteMol;
                     isoValueType: Bootstrap.Visualization.Density.IsoValueType.Sigma,
                     color: LiteMol.Visualization.Color.fromHex(0x3362B2),
                     isWireframe: true,
+                    disableFog: true,
                     transparency: { alpha: 0.75 }
                 })
             }, { ref: 'density-2fo-fc' });