+++ examples/CustomDensity/src/Data.ts
@@ -44,6 +44,7 @@ namespace LiteMol.Custom {
                     isoValueType: Bootstrap.Visualization.Density.IsoValueType.Sigma,
                     color: LiteMol.Visualization.Color.fromHex(0x3362B2), 
                     isWireframe: true,
+                    disableFog: true,
                     transparency: { alpha: 0.75 }
                 })
             }, { ref: 'density-2fo-fc' });