+++ src/Viewer/LiteMol-viewer.js
@@ -100,6 +100,7 @@ var LiteMol;
                             isoValueType: Bootstrap.Visualization.Density.IsoValueType.Sigma,
                             color: LiteMol.Visualization.Color.fromHex(0xBB3333),
                             isWireframe: true,
+                            disableFog: true,
                             transparency: { alpha: 1.0 }
                         })
                     });
@@ -114,6 +115,7 @@ var LiteMol;
                             isoValueType: Bootstrap.Visualization.Density.IsoValueType.Sigma,
                             color: LiteMol.Visualization.Color.fromHex(0x33BB33),
                             isWireframe: true,
+                            disableFog: true,
                             transparency: { alpha: 1.0 }
                         })
                     });
@@ -130,6 +132,7 @@ var LiteMol;
                             isoValueType: Bootstrap.Visualization.Density.IsoValueType.Sigma,
                             color: LiteMol.Visualization.Color.fromHex(0x3362B2),
                             isWireframe: false,
+                            disableFog: true,
                             transparency: { alpha: 0.45 }
                         })
                     });