+++ src/Viewer/PDBe/Data.ts
@@ -79,6 +79,7 @@ namespace LiteMol.Viewer.PDBe.Data {
                     isoValueType: Bootstrap.Visualization.Density.IsoValueType.Sigma,
                     color: LiteMol.Visualization.Color.fromHex(0xBB3333), // can also use fromRgb(0-255 ranges), fromHsl, fromHsv; found in Visualization/Base/Theme.ts
                     isWireframe: true,
+                    disableFog: true,
                     transparency: { alpha: 1.0 }
                 })
             });
@@ -94,6 +95,7 @@ namespace LiteMol.Viewer.PDBe.Data {
                     isoValueType: Bootstrap.Visualization.Density.IsoValueType.Sigma,
                     color: LiteMol.Visualization.Color.fromHex(0x33BB33), 
                     isWireframe: true,
+                    disableFog: true,
                     transparency: { alpha: 1.0 }
                 })
             });            
@@ -111,6 +113,7 @@ namespace LiteMol.Viewer.PDBe.Data {
                     isoValueType: Bootstrap.Visualization.Density.IsoValueType.Sigma,
                     color: LiteMol.Visualization.Color.fromHex(0x3362B2), 
                     isWireframe: false,
+                    disableFog: true,
                     transparency: { alpha: 0.45 }
                 })
             });