+++ web/Examples/CustomControls/LiteMol-example.js
@@ -93,6 +93,7 @@ var LiteMol;
                             isoValueType: Bootstrap.Visualization.Density.IsoValueType.Sigma,
                             color: LiteMol.Visualization.Color.fromHex(0xBB3333),
                             isWireframe: true,
+                            disableFog: true,
                             transparency: { alpha: 1.0 }
                         })
                     });
@@ -107,6 +108,7 @@ var LiteMol;
                             isoValueType: Bootstrap.Visualization.Density.IsoValueType.Sigma,
                             color: LiteMol.Visualization.Color.fromHex(0x33BB33),
                             isWireframe: true,
+                            disableFog: true,
                             transparency: { alpha: 1.0 }
                         })
                     });
@@ -123,6 +125,7 @@ var LiteMol;
                             isoValueType: Bootstrap.Visualization.Density.IsoValueType.Sigma,
                             color: LiteMol.Visualization.Color.fromHex(0x3362B2),
                             isWireframe: false,
+                            disableFog: true,
                             transparency: { alpha: 0.45 }
                         })
                     });