+++ src/Bootstrap/CHANGELOG.md
@@ -1,5 +1,7 @@
 ﻿# 1.2.2
 * Added support for popup messages (toasts).
+* Added the option to disable fog on per visual basis.
+* Fog no longer affects selected residue, ambience, and dynamic density data.
 
 # 1.2.1
 * Transform builder no longer requires verbose type annotations that were sometimes required.