+++ web/Viewer/CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.1.11
 * Added support for popup messages (toasts).
+* Fog no longer affects selected residue, ambience, and dynamic density data.
 
 # 1.1.10
 * Changed how visual interactivity works.